Limit weather to first five levels and outdoor locations.
In spite of all the work I put in on the weather system, in practice
it's not very fun to deal with while dungeon delving, and it also
breaks the sense of immersion many players have. As such, it's going
to become a bit of a vestigal feature, only impacting a very small
number of areas. I'm still interested in expanding it in the future,
but I'll have to think on how.